Fix -fsanitize=integer unsigned left shift errors
Fix -fsanitize=integer runtime errors like:
  left shift of 4294961503 by 16 places cannot be represented in type
  'uint32_t' (aka 'unsigned int')

For b of type int16_t, the expression (uint32_t)(uint16_t)(b) guarantees
that the upper 16 bits of the uint32_t value are all zeros. Then the
result of << 16 can be represented in uint32_t.

+++ aom_dsp/x86/txfm_common_avx2.h
@@ -22,7 +22,7 @@ extern "C" {
 
 static INLINE __m256i pair_set_w16_epi16(int16_t a, int16_t b) {
   return _mm256_set1_epi32(
-      (int32_t)(((uint16_t)(a)) | (((uint32_t)(b)) << 16)));
+      (int32_t)(((uint16_t)(a)) | (((uint32_t)(uint16_t)(b)) << 16)));
 }
 
 static INLINE void btf_16_w16_avx2(const __m256i w0, const __m256i w1,
